helmet/apps/visor-ui/requirements.txt
PySide6==6.6.0
grpcio==1.60.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
psutil==5.9.6